TP PRINCIPAL LABORATORIO DE CONSTRUCCION DE SOFTWARE/Modulo de operaciones crud-Calculo de nomina/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click== 8.1.8
colorama==0.4.6
gunicorn==23.0.0


fastapi==0.115.12
idna==3.10
numpy==1.26.4

pillow==11.2.1
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2

passlib==1.7.4
python-jose==3.3.0
cloudinary==1.37.0
python-multipart==0.0.20
bcrypt==4.3.0
reportlab==4.4.1
python-dotenv==1.1.0

jinja2==3.1.6
weasyprint==65.1
num2words==0.5.14



